paropt-service-sdk/requirements.txt
requests==2.20.0
jsonschema==2.6.0
sphinx_rtd_theme
nbsphinx
configobj
globus_sdk
mdf_toolbox
pyyaml